requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.6
django-annoying==0.9.0
django-dotenv==1.4.1
django-eventtools==0.9.7
django-localflavor==1.3
django-recurrence==1.3.0
django-scheduler==0.8
gnureadline==6.3.3
googlemaps==2.4.3
gunicorn==19.6.0
icalendar==3.9.2
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.3
pytz==2016.4
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
wheel==0.24.0
